Minor code refactoring

+++ Bank Simulator/Models/DatabaseModels.cs
@@ -12,7 +12,7 @@ namespace Bank_Simulator.Models
         public string CardNumber { get; set; }
         public string AccountNumber { get; set; }
         public string ExpiryDate { get; set; }
-        public int AccountBalance { get; set; }
+        public double AccountBalance { get; set; }
         public string CVV { get; set; }
         public string PIN { get; set; }
